requirements.txt
click==6.2
ordereddict==1.1
PyAudio==0.2.9
python-dotenv==0.3.0
requests==2.8.1
watson-developer-cloud==0.2.1
wheel==0.24.0
